restructured tests.py into 4 separate modules with common fixture
