--- blogspot-project/diary/test_api.py
@@ -1,958 +1,0 @@
-from pprint import pprint
-import random
-import re
-from unittest import skip
-from .serializers import (
-    PostSerializer,
-    PostDetailSerializer,
-    LikeSerializer,
-    LikeDetailSerializer,
-    UserSerializer,
-    UserDetailSerializer,
-)
-from .models import CustomUser, Post, Like
-from rest_framework import status
-from rest_framework.reverse import reverse
-from rest_framework.test import APITestCase
-from rest_framework.settings import api_settings
-from django.db.models import Count
-from rest_framework_simplejwt.tokens import RefreshToken
-from django.core.exceptions import FieldError
-from django.core import mail
-
-
-class PostAPITestCase(APITestCase):
-    @classmethod
-    def setUpTestData(cls):
-
-        cls.admin = CustomUser.objects.create_user(
-            email="admin@ukr.net", username="admin", password="fokker123", is_staff=True
-        )
-        cls.test_user_1 = CustomUser.objects.create_user(
-            email="test1@ukr.net", username="TestUser1", password="fokker123"
-        )
-        cls.test_user_2 = CustomUser.objects.create_user(
-            email="test2@ukr.net", username="TestUser2", password="fokker123"
-        )
-        cls.test_user_3 = CustomUser.objects.create_user(
-            email="test3@ukr.net", username="TestUser3", password="fokker123"
-        )
-
-        cls.test_post_1 = Post.objects.create(
-            title="TestPost1", author=cls.test_user_1, content="Some test 1 content"
-        )
-        cls.test_post_2 = Post.objects.create(
-            title="TestPost2", author=cls.test_user_3, content="Some test 2 content"
-        )
-        cls.test_post_3 = Post.objects.create(
-            title="TestPost3",
-            author=cls.test_user_1,
-            content="Some test 3 content",
-            published=False,
-        )
-        cls.test_post_4 = Post.objects.create(
-            title="TestPost4", author=cls.test_user_2, content="Some test 4 content"
-        )
-        cls.test_post_5 = Post.objects.create(
-            title="TestPost5", author=cls.test_user_3, content="Some test 5 content"
-        )
-        cls.test_post_6 = Post.objects.create(
-            title="TestPost6", author=cls.test_user_3, content="Some test 6 content"
-        )
-        cls.test_post_7 = Post.objects.create(
-            title="TestPost7", author=cls.test_user_1, content="Some test 7 content"
-        )
-        cls.test_post_8 = Post.objects.create(
-            title="TestPost8", author=cls.test_user_2, content="Some test 8 content"
-        )
-        cls.test_post_9 = Post.objects.create(
-            title="TestPost9", author=cls.test_user_3, content="Some test 9 content"
-        )
-        cls.test_post_10 = Post.objects.create(
-            title="TestPost10", author=cls.test_user_2, content="Some test 10 content"
-        )
-        cls.test_post_11 = Post.objects.create(
-            title="TestPost11", author=cls.test_user_3, content="Some test 11 content"
-        )
-        cls.test_post_12 = Post.objects.create(
-            title="TestPost12",
-            author=cls.test_user_3,
-            content="Some test 11 content",
-            published=False,
-        )
-
-        # Obtaining an access token for further authorization by JWT-tokens.
-        cls.access_token_admin = RefreshToken.for_user(cls.admin).access_token
-        cls.access_token_user1 = RefreshToken.for_user(cls.test_user_1).access_token
-        cls.access_token_user2 = RefreshToken.for_user(cls.test_user_2).access_token
-        cls.access_token_user3 = RefreshToken.for_user(cls.test_user_3).access_token
-
-        cls.test_like1 = Like.objects.create(user=cls.test_user_1, post=cls.test_post_1)
-        cls.test_like2 = Like.objects.create(user=cls.test_user_1, post=cls.test_post_3)
-        cls.test_like3 = Like.objects.create(user=cls.test_user_1, post=cls.test_post_5)
-        cls.test_like4 = Like.objects.create(user=cls.test_user_1, post=cls.test_post_9)
-        cls.test_like5 = Like.objects.create(user=cls.test_user_1, post=cls.test_post_7)
-        cls.test_like6 = Like.objects.create(user=cls.test_user_1, post=cls.test_post_2)
-        cls.test_like7 = Like.objects.create(user=cls.test_user_2, post=cls.test_post_1)
-        cls.test_like8 = Like.objects.create(user=cls.test_user_2, post=cls.test_post_2)
-        cls.test_like9 = Like.objects.create(user=cls.test_user_2, post=cls.test_post_3)
-        cls.test_like10 = Like.objects.create(
-            user=cls.test_user_2, post=cls.test_post_8
-        )
-        cls.test_like11 = Like.objects.create(
-            user=cls.test_user_2, post=cls.test_post_10
-        )
-        cls.test_like12 = Like.objects.create(
-            user=cls.test_user_2, post=cls.test_post_11
-        )
-        cls.test_like13 = Like.objects.create(
-            user=cls.test_user_2, post=cls.test_post_5
-        )
-        cls.test_like14 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_11
-        )
-        cls.test_like15 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_3
-        )
-        cls.test_like16 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_2
-        )
-        cls.test_like17 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_1
-        )
-        cls.test_like18 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_4
-        )
-        cls.test_like19 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_7
-        )
-        cls.test_like20 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_8
-        )
-        cls.test_like21 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_9
-        )
-        cls.test_like22 = Like.objects.create(
-            user=cls.test_user_3, post=cls.test_post_6
-        )
-
-    def test_post_list(self):
-
-        queryset = (
-            Post.objects.exclude(published=False)
-            .annotate(likes=Count("like"))
-            .order_by("-updated")
-        )
-
-        response = self.client.get(reverse("post-list-create-api"))
-
-        serializer = PostSerializer(
-            queryset, many=True, context={"request": response.wsgi_request}
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            serializer.data[: api_settings.PAGE_SIZE], response.data["results"]
-        )
-
-    def test_post_create(self):
-
-        # We just need to include Authorization header with our post-request below.
-        # We also can use .credentials to set Authorization header with all requests:
-        # self.client.credentials(HTTP_AUTHORIZATION=f"JWT {access_token}")
-        # but we need 1 request to be Unauthorized, so we won't do it for now.
-        # Or we can use standard Django session authorization backend:
-        # self.client.login(username="TestUser1", password="fokker123")
-
-        # Unauthorized
-        response1 = self.client.post(
-            reverse("post-list-create-api"),
-            {"title": "New Test Post 1", "content": "Some conntent of New Test Post 1"},
-        )
-
-        # Use credentials to set Authorization header with all further requests:
-        self.client.credentials(HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}")
-
-        # Bad request. Lack of "title".
-        response2 = self.client.post(
-            reverse("post-list-create-api"),
-            {"content": "Some conntent of New Test Post 2"},
-        )
-
-        response3 = self.client.post(
-            reverse("post-list-create-api"),
-            {"title": "New Test Post 3", "content": "Some conntent of New Test Post 3"},
-        )
-
-        response4 = self.client.post(
-            reverse("post-list-create-api"),
-            {
-                "title": "New Test Post 4",
-                "content": "Some conntent of New Test Post 4",
-                "author": 3,
-            },
-        )
-
-        response5 = self.client.post(
-            reverse("post-list-create-api"),
-            {
-                "title": "New Test Post 5",
-                "content": "Some conntent of New Test Post 5",
-                "created": "2022-03-01",
-            },
-        )
-
-        response6 = self.client.post(
-            reverse("post-list-create-api"),
-            {
-                "title": "New Test Post 6",
-                "content": "Some conntent of New Test Post 6",
-                "published": False,
-            },
-        )
-
-        serializer = PostSerializer(
-            Post.objects.get(title="New Test Post 3"),
-            context={"request": response1.wsgi_request},
-        )
-
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertRaises(Post.DoesNotExist, Post.objects.get, title="New Test Post 1")
-
-        self.assertEqual(response2.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertRaises(Post.DoesNotExist, Post.objects.get, title="New Test Post 2")
-
-        self.assertEqual(response3.status_code, status.HTTP_201_CREATED)
-        self.assertTrue(Post.objects.get(title="New Test Post 3"))
-        self.assertEqual(serializer.data, response3.data)
-
-        self.assertEqual(response4.status_code, status.HTTP_201_CREATED)
-        self.assertNotEqual(Post.objects.get(title="New Test Post 4").author_id, 3)
-        self.assertEqual(
-            Post.objects.get(title="New Test Post 4").author_id,
-            response4.wsgi_request.user.id,
-        )
-
-        self.assertEqual(response5.status_code, status.HTTP_201_CREATED)
-        self.assertNotEqual(
-            Post.objects.get(title="New Test Post 5").created, "2022-03-01"
-        )
-
-        self.assertEqual(response6.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(Post.objects.get(title="New Test Post 6").published, False)
-
-    def test_post_detail(self):
-
-        # It's good idea to make a random choosing of Post object,
-        # but we need to test both Post object with published Ture and False.
-        # post_id = random.choice(Post.objects.all().values('id'))['id']
-        # url = (reverse("post-detail-api", args=[post_id]))
-
-        # published=True
-        response1 = self.client.get(
-            reverse("post-detail-api", args=[self.test_post_1.id])
-        )
-        # published=False
-        response2 = self.client.get(
-            reverse("post-detail-api", args=[self.test_post_12.id])
-        )
-
-        # published=False, but read by Owner
-        response3 = self.client.get(
-            reverse("post-detail-api", args=[self.test_post_12.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user3}",
-        )
-
-        serializer1 = PostDetailSerializer(
-            self.test_post_1,
-            context={"request": response1.wsgi_request},
-        )
-
-        serializer3 = PostDetailSerializer(
-            self.test_post_12,
-            context={"request": response1.wsgi_request},
-        )
-
-        self.assertEqual(response1.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer1.data, response1.data)
-
-        self.assertEqual(response2.status_code, status.HTTP_403_FORBIDDEN)
-
-        self.assertEqual(response3.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer3.data, response3.data)
-
-    def test_post_update(self):
-        """
-        IMHO.
-        It's worth nothing that Post objects reachable with self.test_post_*
-        only accessibe for retrieving so they don't change their state.
-        In case we need to see wether the object is changed we should retrieve it from the object manager.
-        """
-
-        # Make a response just to obtain the wsgi_request for serializer to get serializer of clear Post object
-        response = self.client.get("/")
-        serializer_before_update = PostDetailSerializer(
-            self.test_post_1, context={"request": response.wsgi_request}
-        )
-
-        # Unathorized
-        response1 = self.client.put(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by put ",
-                "content": "Some UPDATED test 1 content",
-            },
-        )
-        serializer_after_update1 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response1.wsgi_request},
-        )
-
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertEqual(serializer_before_update.data, serializer_after_update1.data)
-
-        # Authorized by owner
-        response2 = self.client.put(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by put",
-                "content": "Some UPDATED test 1 content",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update2 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response2.wsgi_request},
-        )
-        self.assertEqual(response2.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update1.data, serializer_after_update2.data
-        )
-
-        # Authorized by admin
-        response3 = self.client.put(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by put",
-                "content": "Some UPDATED test 1 content",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        serializer_after_update3 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response3.wsgi_request},
-        )
-        self.assertEqual(response3.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update2.data, serializer_after_update3.data
-        )
-
-        # Update with unexisted fieldname
-        response4 = self.client.put(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by put",
-                "contAnt": "Some UPDATED test 1 content",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update4 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response4.wsgi_request},
-        )
-        self.assertEqual(response4.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            serializer_after_update3.data["content"],
-            serializer_after_update4.data["content"],
-        )
-
-        ##############################PATCH
-        # Unathorized
-        response1 = self.client.patch(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by patch ",
-                "content": "Some UPDATED with patch test 1 content",
-            },
-        )
-        serializer_after_update5 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response1.wsgi_request},
-        )
-
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertEqual(serializer_after_update4.data, serializer_after_update5.data)
-
-        # Authorized by owner
-        response2 = self.client.patch(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by patch 2",
-                "content": "Some UPDATED with patch 2 test 1 content",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update6 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response2.wsgi_request},
-        )
-        self.assertEqual(response2.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update5.data, serializer_after_update6.data
-        )
-
-        # Authorized by admin
-        response3 = self.client.patch(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by patch 3",
-                "content": "Some UPDATED with patch 3 test 1 content",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        serializer_after_update7 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response3.wsgi_request},
-        )
-        self.assertEqual(response3.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update6.data, serializer_after_update7.data
-        )
-
-        # Update with unexisted fieldname
-        response4 = self.client.patch(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            {
-                "title": "TestPost1 UPDATED by patch 4",
-                "contAnt": "Some UPDATED with patch 4 test 1 content",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update8 = PostDetailSerializer(
-            Post.objects.get(id=self.test_post_1.id),
-            context={"request": response4.wsgi_request},
-        )
-        self.assertEqual(response4.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            serializer_after_update7.data["content"],
-            serializer_after_update8.data["content"],
-        )
-
-    def test_post_delete(self):
-
-        # Unauthorized
-        response1 = self.client.delete(
-            reverse("post-detail-api", args=[self.test_post_1.id])
-        )
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertTrue(Post.objects.get(id=self.test_post_1.id))
-
-        # Authorized by owner
-        response2 = self.client.delete(
-            reverse("post-detail-api", args=[self.test_post_1.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        self.assertEqual(response2.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertRaises(Post.DoesNotExist, Post.objects.get, title="TestPost1")
-
-        # Authorized by admin
-        response2 = self.client.delete(
-            reverse("post-detail-api", args=[self.test_post_2.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        self.assertEqual(response2.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertRaises(Post.DoesNotExist, Post.objects.get, title="TestPost2")
-
-    def test_like_list(self):
-
-        queryset = (
-            Like.objects.values("created__date")
-            .annotate(likes=Count("id"))
-            .order_by("-created__date")
-        )
-
-        response = self.client.get(reverse("like-list-api"))
-
-        serializer = LikeSerializer(
-            queryset, many=True, context={"request": response.wsgi_request}
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer.data, response.data["results"])
-
-    def test_like_detail(self):
-
-        response = self.client.get(
-            reverse("like-detail-api", args=[self.test_like1.id])
-        )
-
-        serializer = LikeDetailSerializer(
-            self.test_like1,
-            context={"request": response.wsgi_request},
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer.data, response.data)
-
-    def test_like_create_delete(self):
-
-        count = self.test_post_11.like_set.count()
-
-        # Unauthorized
-        response1 = self.client.get(
-            reverse("like-create-api", args=[self.test_post_11.id])
-        )
-
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertEqual(count, self.test_post_11.like_set.count())
-
-        # Authorized by unliked user = CREATE the like
-        response2 = self.client.get(
-            reverse("like-create-api", args=[self.test_post_11.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-
-        self.assertEqual(response2.status_code, status.HTTP_201_CREATED)
-        self.assertEqual(count + 1, self.test_post_11.like_set.count())
-
-        # Authorized by liked user = DELETE the like
-        response3 = self.client.get(
-            reverse("like-create-api", args=[self.test_post_11.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-
-        self.assertEqual(response3.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertEqual(count, self.test_post_11.like_set.count())
-
-        # Post doesn't exist
-        response3 = self.client.get(
-            reverse("like-create-api", args=[48]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-
-        self.assertEqual(response3.status_code, status.HTTP_404_NOT_FOUND)
-        self.assertEqual(count, self.test_post_11.like_set.count())
-
-    def test_user_list(self):
-
-        queryset = CustomUser.objects.all().order_by("-last_request")
-
-        # Unauthorized
-        response1 = self.client.get(reverse("user-list-create-api"))
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-
-        # Authorized as non-staff user
-        response2 = self.client.get(
-            reverse("user-list-create-api"),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        self.assertEqual(response2.status_code, status.HTTP_403_FORBIDDEN)
-
-        # Authorized as admin
-        response3 = self.client.get(
-            reverse("user-list-create-api"),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        serializer = UserSerializer(
-            queryset, many=True, context={"request": response3.wsgi_request}
-        )
-        self.assertEqual(response3.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer.data, response3.data["results"])
-
-    def test_user_create(self):
-
-        # Authorized
-        response1 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser",
-                "email": "somemail@gmail.com",
-                "password": "ribark8903",
-                "password2": "ribark8903",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        self.assertEqual(response1.status_code, status.HTTP_403_FORBIDDEN)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="NewTestUser"
-        )
-
-        # Unauthorized, correct data
-        response2 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser",
-                "email": "somemail@gmail.com",
-                "password": "ribark8903",
-                "password2": "ribark8903",
-            },
-        )
-        serializer = UserSerializer(
-            CustomUser.objects.get(username="NewTestUser"),
-            context={"request": response2.wsgi_request},
-        )
-        self.assertEqual(response2.status_code, status.HTTP_201_CREATED)
-        self.assertTrue(CustomUser.objects.get(username="NewTestUser"))
-        self.assertEqual(serializer.data, response2.data)
-
-        # Password2 missed
-        response3 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser2",
-                "email": "somemail2@gmail.com",
-                "password": "ribark8903cz",
-            },
-        )
-        self.assertEqual(response3.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="NewTestUser2"
-        )
-
-        # Email missed
-        response4 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser2",
-                "password": "ribark8903cz",
-                "password2": "ribark8903cz",
-            },
-        )
-        self.assertEqual(response4.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="NewTestUser2"
-        )
-
-        # Invalid email
-        response5 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser2",
-                "email": "sdncsja.io",
-                "password": "ribark8903cz",
-                "password2": "ribark8903cz",
-            },
-        )
-        self.assertEqual(response5.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="NewTestUser2"
-        )
-
-        # Passwords doesn't match
-        response6 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser2",
-                "email": "sdncsja.io",
-                "password": "ribark8903cz",
-                "password2": "ribark8903",
-            },
-        )
-        self.assertEqual(response6.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="NewTestUser2"
-        )
-
-        # Passwords similar with username
-        response7 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser2",
-                "email": "somemail2@gmail.com",
-                "password": "newtestuser2",
-                "password2": "newtestuser2",
-            },
-        )
-        self.assertEqual(response7.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="NewTestUser2"
-        )
-
-        # Unnessesary fields
-        response8 = self.client.post(
-            reverse("user-list-create-api"),
-            {
-                "username": "NewTestUser2",
-                "email": "somemail2@gmail.com",
-                "password": "ribark8903cz",
-                "password2": "ribark8903cz",
-                "sex": "Male",
-            },
-        )
-        object = CustomUser.objects.get(username="NewTestUser2")
-        serializer8 = UserSerializer(
-            object, context={"request": response2.wsgi_request}
-        )
-        self.assertEqual(response2.status_code, status.HTTP_201_CREATED)
-        self.assertFalse(getattr(object, "sex", False))
-        self.assertRaises(FieldError, CustomUser.objects.get, sex="Male")
-        self.assertEqual(serializer8.data, response8.data)
-
-    def test_user_detail(self):
-
-        # Authorized by owner
-        response = self.client.get(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer = UserDetailSerializer(
-            self.test_user_1,
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer.data, response.data)
-
-        # Authorized by admin
-        response = self.client.get(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        serializer = UserDetailSerializer(
-            self.test_user_1,
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(serializer.data, response.data)
-
-        # Authorized by non-owner (and non-admin)
-        response = self.client.get(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user2}",
-        )
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-        # Unauthorized
-        response = self.client.get(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-        )
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-
-    def test_user_update(self):
-        """
-        IMHO.
-        It's worth nothing that Post objects reachable with self.test_post_*
-        only accessibe for retrieving so they don't change their state.
-        In case we need to see wether the object is changed we should retrieve it from the object manager.
-        """
-
-        # Make a response just to obtain the wsgi_request for serializer to get serializer of clear Post object
-        response_initial = self.client.get("/")
-        serializer_before_update = UserDetailSerializer(
-            self.test_user_1, context={"request": response_initial.wsgi_request}
-        )
-
-        # Unathorized
-        response = self.client.put(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            {
-                "username": "TestUser1 UPDATED by put",
-                "email": "newemail@ukr.net",
-            },
-        )
-        serializer_after_update1 = UserDetailSerializer(
-            CustomUser.objects.get(id=self.test_user_1.id),
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertEqual(serializer_before_update.data, serializer_after_update1.data)
-
-        # Authorized by owner
-        response = self.client.put(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            {
-                "username": "TestUser1PUTed",
-                "email": "newemail@ukr.net",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update2 = UserDetailSerializer(
-            CustomUser.objects.get(id=self.test_user_1.id),
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update1.data, serializer_after_update2.data
-        )
-        self.assertEqual(serializer_after_update2.data["username"], "TestUser1PUTed")
-        self.assertEqual(serializer_after_update2.data["email"], "newemail@ukr.net")
-
-        # Authorized by admin
-        response = self.client.put(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            {
-                "username": "TestUser1PUTedbyAdmin",
-                "email": "newemailadm@ukr.net",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        serializer_after_update3 = UserDetailSerializer(
-            CustomUser.objects.get(id=self.test_user_1.id),
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update2.data, serializer_after_update3.data
-        )
-        self.assertEqual(
-            serializer_after_update3.data["username"], "TestUser1PUTedbyAdmin"
-        )
-        self.assertEqual(serializer_after_update3.data["email"], "newemailadm@ukr.net")
-
-        ######PATCH
-        # Authorized by owner, incorrect email
-        response = self.client.patch(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            {"username": "TestUser1PATCHed", "email": "asdnaa1223"},
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update4 = UserDetailSerializer(
-            CustomUser.objects.get(id=self.test_user_1.id),
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(serializer_after_update3.data, serializer_after_update4.data)
-
-        # Authorized by owner, unnessesary unknown field
-        response = self.client.patch(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            {
-                "username": "TestUser1PATCHed",
-                "email": "asdnaa1223@gmail.com",
-                "sex": "Female",
-            },
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        serializer_after_update5 = UserDetailSerializer(
-            CustomUser.objects.get(id=self.test_user_1.id),
-            context={"request": response.wsgi_request},
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            serializer_after_update4.data, serializer_after_update5.data
-        )
-        self.assertEqual(serializer_after_update5.data["username"], "TestUser1PATCHed")
-        self.assertEqual(serializer_after_update5.data["email"], "asdnaa1223@gmail.com")
-        self.assertRaises(FieldError, CustomUser.objects.get, sex="Female")
-
-        # PASSWORD change by owner
-        response = self.client.patch(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_1.id]),
-            {"password": "fokker1234"},
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user1}",
-        )
-        # Please recall that self.test_user_1.password reffered to the virgin user on the test start
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertNotEqual(
-            CustomUser.objects.get(id=self.test_user_1.id).password,
-            self.test_user_1.password,
-        )
-
-    def test_user_delete(self):
-        # Unauthorized
-        response = self.client.delete(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_2.id])
-        )
-        self.assertEqual(response.status_code, status.HTTP_401_UNAUTHORIZED)
-        self.assertTrue(CustomUser.objects.get(id=self.test_user_2.id))
-
-        # Authorized by owner
-        response = self.client.delete(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_2.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_user2}",
-        )
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="TestUser2"
-        )
-
-        # Authorized by admin
-        response = self.client.delete(
-            reverse("user-detail-update-destroy-api", args=[self.test_user_3.id]),
-            HTTP_AUTHORIZATION=f"JWT {self.access_token_admin}",
-        )
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
-        self.assertRaises(
-            CustomUser.DoesNotExist, CustomUser.objects.get, username="TestUser3"
-        )
-
-    def test_jwt_authentication(self):
-        # Create new user
-        test_user_4 = CustomUser.objects.create_user(
-            email="testuser4@ukr.net", username="TestUser4", password="fokker12345"
-        )
-
-        # Login-api icorrect password
-        response1 = self.client.post(
-            reverse("login-api"),
-            {
-                "username": test_user_4.username,
-                "password": "Kaskmcs12341",
-            },
-        )
-        self.assertEqual(response1.status_code, status.HTTP_401_UNAUTHORIZED)
-
-        # Login-api correct data
-        response2 = self.client.post(
-            reverse("login-api"),
-            {
-                "username": test_user_4.username,
-                "password": "fokker12345",
-            },
-        )
-        self.assertEqual(response2.status_code, status.HTTP_200_OK)
-
-        # Verify that refresh token (and access token too) is valid
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": response2.data["refresh"]}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_200_OK)
-
-        # incorrect token
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": "jsndvkajsdnvkajsnv"}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_401_UNAUTHORIZED)
-
-        # Token-refresh-api
-        response3 = self.client.post(
-            reverse("token-refresh-api"),
-            {
-                "refresh": response2.data["refresh"],
-            },
-        )
-        self.assertEqual(response3.status_code, status.HTTP_200_OK)
-        self.assertTrue(response3.data["access"])
-        self.assertTrue(response3.data["refresh"])
-        # Check wether new refreshed refresh token is valid
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": response3.data["refresh"]}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_200_OK)
-        # Check wether OLD refresh token is invalid
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": response2.data["refresh"]}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_400_BAD_REQUEST)
-
-        # Token-recovery-api
-        response = self.client.post(
-            reverse("token-recovery-api"), {"email": test_user_4.email}
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        # Get the new 'access' token from email which has been emailed to the user
-        # It will be use by user to change password
-        new_access_token = re.search(r"(?<='access': ).+", mail.outbox[0].body)
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": new_access_token.group()}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_200_OK)
-        # Check wether OLD REFRESHED refresh token is invalid
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": response3.data["refresh"]}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_400_BAD_REQUEST)
-        # Check wether OLD-OLD refresh token is invalid
-        verify_token = self.client.post(
-            reverse("token_verify"), {"token": response2.data["refresh"]}
-        )
-        self.assertEqual(verify_token.status_code, status.HTTP_400_BAD_REQUEST)
